+++ input/zmq_input.cpp
@@ -4,7 +4,7 @@
 #include "utilities/Logs.h"
 #include "simulation/simulation.h"
 #include "simulation/simulationtime.h"
-#include "vehicle/Train.h"
+#include "vehicle/train/Train.h"
 
 zmq_input::zmq_input()
 {
