+++ vehicle/train/TrainCommandsControls.cpp
@@ -8,7 +8,7 @@ http://mozilla.org/MPL/2.0/.
 */
 
 #include "stdafx.h"
-#include "vehicle/Train.h"
+#include "vehicle/train/Train.h"
 
 #include "world/Event.h"
 #include "simulation/simulationtime.h"
