+++ vehicle/train/TrainCommandsLights.cpp
@@ -8,7 +8,7 @@ http://mozilla.org/MPL/2.0/.
 */
 
 #include "stdafx.h"
-#include "vehicle/Train.h"
+#include "vehicle/train/Train.h"
 
 #include "simulation/simulation.h"
 #include "world/Event.h"
